ENH: Fixed mergeTraces so that there are fewer errors in automated joining

+++ Tracing/TraceEdit/View3D.h
@@ -122,7 +122,7 @@ protected:
   void Rerender();
 
 private:
-	int gapTol;
+	double gapTol;
 	int gapMax;
 	int smallLine;
 	float lineWidth;
